This should complete the first part of my journey of running Mir on
[Chimera Linux](https://chimera-linux.org) initially reported on #3383.

Now I'm figuring out this runtime failure of `mir_demo_server`
(https://paste.c-net.org/umtbetmnnvsn) which I cannot reproduce on Arch
with (https://paste.c-net.org/a3azgse7x8en) or without these patches,
further opinions welcome :)

+++ src/server/console/minimal_console_services.cpp
@@ -100,7 +100,7 @@ std::future<std::unique_ptr<mir::Device>> mir::MinimalConsoleServices::acquire_d
     auto const devnode = [](auto const& fd) {
         using namespace boost::iostreams;
         char line_buffer[1024];
-        stream<file_descriptor_source> uevent{fd, file_descriptor_flags::never_close_handle};
+        stream<file_descriptor_source> uevent{static_cast<int>(fd), file_descriptor_flags::never_close_handle};
 
         while (uevent.getline(line_buffer, sizeof(line_buffer)))
         {
